hotfix rollback tables

+++ database/migrations/2024_01_30_155316_add_new_field_to_dur_table.php
@@ -21,8 +21,10 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::table('dur', function (Blueprint $table) {
-            $table->dropColumn('applicant_id');
-        });
+        if (Schema::hasTable('dur')) {
+            Schema::table('dur', function (Blueprint $table) {
+                $table->dropColumn('applicant_id');
+            });
+        }
     }
 };
